LUT-33076 : Add RBAC control to delete questions in form

+++ src/java/fr/paris/lutece/plugins/forms/web/admin/FormQuestionJspBean.java
@@ -61,6 +61,7 @@ import fr.paris.lutece.plugins.forms.business.StepHome;
 import fr.paris.lutece.plugins.forms.service.FormDatabaseService;
 import fr.paris.lutece.plugins.forms.service.FormDisplayService;
 import fr.paris.lutece.plugins.forms.service.FormService;
+import fr.paris.lutece.plugins.forms.service.FormsResourceIdService;
 import fr.paris.lutece.plugins.forms.service.IFormDatabaseService;
 import fr.paris.lutece.plugins.forms.service.IFormDisplayService;
 import fr.paris.lutece.plugins.forms.util.FormsConstants;
@@ -73,9 +74,7 @@ import fr.paris.lutece.plugins.genericattributes.business.Field;
 import fr.paris.lutece.plugins.genericattributes.business.FieldHome;
 import fr.paris.lutece.plugins.genericattributes.service.entrytype.EntryTypeServiceManager;
 import fr.paris.lutece.plugins.genericattributes.service.entrytype.IEntryTypeService;
-import fr.paris.lutece.portal.business.user.AdminUser;
 import fr.paris.lutece.portal.service.admin.AccessDeniedException;
-import fr.paris.lutece.portal.service.admin.AdminUserService;
 import fr.paris.lutece.portal.service.file.FileServiceException;
 import fr.paris.lutece.portal.service.image.ImageResourceManager;
 import fr.paris.lutece.portal.service.message.AdminMessage;
@@ -127,7 +126,7 @@ public class FormQuestionJspBean extends AbstractFormQuestionJspBean
     // Warning messages
     private static final String WARNING_CONFIRM_REMOVE_QUESTION_FORM_ACTIVE = "forms.warning.deleteComposite.confirmRemoveQuestion.formActive";
     private static final String WARNING_CONFIRM_REMOVE_GROUP_ANY_QUESTIONS_FORM_ACTIVE = "forms.warning.deleteComposite.confirmRemoveGroup.formActive";
-    private static final String WARNING_INSUFFICIENT_RIGHTS_LEVEL_TO_DELETE_COMPOSITE = "forms.warning.deleteComposite.insufficientRightsLevel";
+    private static final String WARNING_INSUFFICIENT_MODIFY_PERMISSION_TO_DELETE_COMPOSITE = "forms.warning.deleteComposite.insufficientModifyPermission";
 
     // Markers
     private static final String MARK_FIELDS_LIST_BY_ID_ENTRIES = "fields_list_by_id_entries";
@@ -135,7 +134,6 @@ public class FormQuestionJspBean extends AbstractFormQuestionJspBean
     // Constants
     private static final String PUBLIC_IMAGE_RESOURCE = "public_image_resource";
     private static final String ILLUSTRATION_IMAGE = "illustration_image";
-    private static final int TECHNICAL_ADMIN_RIGHT_LEVEL = 0;
     
     // Other
     @Inject
@@ -576,12 +574,17 @@ public class FormQuestionJspBean extends AbstractFormQuestionJspBean
             _formDisplay = FormDisplayHome.findByPrimaryKey( nIdDisplay );
         }
 
-        AdminUser adminUser = AdminUserService.getAdminUser( request );
-        if ( adminUser != null && adminUser.getUserLevel( ) != TECHNICAL_ADMIN_RIGHT_LEVEL )
+        try
         {
-        	return redirect( request, AdminMessageService.getMessageUrl( request, WARNING_INSUFFICIENT_RIGHTS_LEVEL_TO_DELETE_COMPOSITE, AdminMessage.TYPE_STOP ) );
+            checkUserPermission( Form.RESOURCE_TYPE, String.valueOf( _form.getId( ) ), FormsResourceIdService.PERMISSION_MODIFY, request, null );
         }
-        else if ( _formService.existCompositeResponses( _formDisplay ) )
+        catch ( AccessDeniedException e )
+        {
+            return redirect( request,
+                    AdminMessageService.getMessageUrl( request, WARNING_INSUFFICIENT_MODIFY_PERMISSION_TO_DELETE_COMPOSITE, AdminMessage.TYPE_STOP ) );
+        }
+
+        if ( _formService.existCompositeResponses( _formDisplay ) )
         {
             boolean bIsQuestion = CompositeDisplayType.QUESTION.getLabel( ).equalsIgnoreCase( _formDisplay.getCompositeType( ) );
             String strMessage = bIsQuestion ? MESSAGE_CANT_REMOVE_ENTRY_RESOURCES_ATTACHED : MESSAGE_CANT_REMOVE_GROUP_RESOURCES_ATTACHED;
@@ -662,7 +665,7 @@ public class FormQuestionJspBean extends AbstractFormQuestionJspBean
      * @return The URL to go after performing the action
      */
     @Action( ACTION_REMOVE_COMPOSITE )
-    public String doRemoveComposite( HttpServletRequest request )
+    public String doRemoveComposite( HttpServletRequest request ) throws AccessDeniedException
     {
         int nIdDisplay = NumberUtils.toInt( request.getParameter( FormsConstants.PARAMETER_ID_DISPLAY ), -1 );
         if ( _formDisplay == null || _formDisplay.getId( ) != nIdDisplay )
@@ -672,8 +675,11 @@ public class FormQuestionJspBean extends AbstractFormQuestionJspBean
 
         if ( _formDisplay == null )
         {
-            redirectToViewManageForm( request );
+            return redirectToViewManageForm( request );
         }
+
+        checkUserPermission( Form.RESOURCE_TYPE, String.valueOf( _formDisplay.getFormId( ) ), FormsResourceIdService.PERMISSION_MODIFY,
+                request, null );
 
         getFormDisplayService( ).deleteDisplayAndDescendants( nIdDisplay );
